feat: add wallet button component
This component is our own implementation of @solana/wallet-adapter-react-ui's WalletMultiButton. We've chosen to do our own implementation because the react-ui component has an opinionated stylesheet included that was a pain to override.
This way we also ensure that the styling of our button stays consistent while still making use of the styled wallet picker modal.